+++ changelog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 - Added names for the keys `rfkill` and `micmute` (#883).
   If you previously used the buttons `missing247` and `missing248`, please update to the new names.
 - Added more MacOS keys (#936)
+- Added keycodes above 255. If you are on linux you can use them now. (#935)
 
 ### Changed